+++ UiMain.py
@@ -11,26 +11,13 @@ from anki.misc.lanes import BaseLane
 
 import Design
 from VehicleControlWindow import vehicleControler
+from helpers import *
 
 try:
     from .VisMapGenerator import generate, flip_h, Vismap, Element
 except ImportError:
     from VisMapGenerator import generate, flip_h, Vismap
 
-def relative_to_file(relpath: str) -> str:
-    return os.path.join(
-        os.path.dirname(os.path.abspath(__file__)),
-        relpath
-    )
-
-def load_image(relpath: str):
-    return pygame.image.load(relative_to_file(relpath))
-
-def rotateSurf(surf: pygame.Surface, orientation: tuple[int,int],addition:int=0) -> pygame.Surface:
-    return pygame.transform.rotate(
-        surf,
-        math.degrees(math.atan2(-orientation[1],orientation[0]))+addition
-    )
 
 class Ui:    
     def __init__(self, vehicles: list[anki.Vehicle],
